% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_layers.R
\name{available_datasets}
\alias{available_datasets}
\title{Available Datasets}
\usage{
available_datasets()
}
\arguments{
\item{layer_vector}{a vector of the layers}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
layer_dataset <- c("annotation_layers","epigenomic","G_E","genomic","transcriptomic")
available_datasets(layer_dataset)
}
