% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide x by y}
\usage{
divide(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value (cannot be zero).}
}
\value{
The result of x divided by y.
}
\description{
Divide x by y
}
\examples{
divide(10, 2)
}
