% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_layers.R
\name{available_layers}
\alias{available_layers}
\title{Get available multiomic layers}
\usage{
available_layers()
}
\value{
Get the list of available multiomic layers
}
\description{
Get available multiomic layers
}
\examples{
available_layers()
}
